requirements.txt
google-genai
yt-dlp
tqdm
python-dotenv
pyyaml
flask
feedparser
requests
reportlab
python-docx
notion-client
